Can update pages in db via BookEditor
