+++ src/components/student/index.js
@@ -2,5 +2,6 @@ import BookEditor from "./_BookEditor";
 import Bookshelf from "./Bookshelf";
 import EditorBookInfo from "./EditorBookInfo";
 import StudentView from "./StudentView";
+import SaveProgressButton from "./SaveProgressButton";
 
-export { BookEditor, Bookshelf, EditorBookInfo, StudentView };
+export { BookEditor, Bookshelf, EditorBookInfo, StudentView, SaveProgressButton };
